Update README.md
I believe that the book that I put "Showing Up for Life" by Bill Gates Sr. should be on this list because I truly believe that this book is awe inspiring.Even though this book might not have the best rating on goodreads when I first read this book in High School it truly touched me on a personal level. The book is about Bill Gates Sr. grew up in the Great depression and how his upbringing began for himself and his children and how we should always take any opportunities we have no matter how big or small.